# (WIP) [EN Performance] Reduce memory use and speedup reading single register