--- a/ledger/complete/wal/compactor_test.go
+++ b/ledger/complete/wal/compactor_test.go
@@ -111,7 +111,7 @@ func Test_Compactor(t *testing.T) {
 			select {
 			case <-co.done:
 				// continue
-			case <-time.After(60 * time.Second):
+			case <-time.After(120 * time.Second):
 				assert.FailNow(t, "timed out")
 			}