### Changes

To reduce memory used when reading single register, make the execution state:
- use a new type `ledger.KeyID` to compute `ledger.Path`.
- use `KeyID` which is a simplified `ledger.Key` with fewer allocs.
- create `ledger.TrieRead` and `ledger.TrieUpdate` without intermediate objects such as `ledger.Query` and `ledger.Update`.

With these changes, `ledger.Key` is only created when updating trie with new payloads.

Updates https://github.com/onflow/flow-go/issues/1744
Updates #2248
Closes #2559 

### Benchmark Comparisons

```
name                     old time/op    new time/op    delta
ExecutionStateRead1-4    7.16µs ± 1%    6.66µs ± 1%   -6.96%  (p=0.000 n=9+10)

name                     old alloc/op   new alloc/op   delta
ExecutionStateRead1-4    1.36kB ± 0%    1.16kB ± 0%  -14.98%  (p=0.000 n=10+8)

name                     old allocs/op  new allocs/op  delta
ExecutionStateRead1-4      15.0 ± 0%      11.0 ± 0%  -26.67%  (p=0.000 n=10+10)

Benchmarks used Go 1.17 on linux_amd64 (Haswell)
```