@fxamacker 2022-12-02
@Kay-Zee  This PR was a POC (for discussion).  I think it's worthwhile to look into other approaches before proceeding with this PR's approach.  I'm closing this PR for now and can open new PR after wrapping up CCF specification and CCF codec.

cc @ramtinms